Move webhooks content
