--- src/KhsCI/Service/Webhooks/Coding.php
@@ -1,16 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace KhsCI\Service\Webhooks;
-
-class Coding
-{
-    public function __construct()
-    {
-    }
-
-    public function __invoke(): void
-    {
-    }
-}
